+++ src/main/kotlin/io/github/plume/oss/domain/files/PythonFile.kt
@@ -1,6 +1,0 @@
-package io.github.plume.oss.domain.files
-
-/**
- * Class wrapper for Python source files.
- */
-class PythonFile(pathname: String) : PlumeFile(pathname)