+++ src/test/resources/intraprocedural/basic/basic5/Basic5.java
@@ -1,4 +1,4 @@
-package intraprocedural.basic.basic6;
+package intraprocedural.basic.basic5;
 
 public class Basic5 {